Crazy amount of loops for a unix time to string conversion

+++ src/handler/net_convert.h
@@ -20,7 +20,7 @@ namespace PDM {
     static int convert(const json &j,NoteMsg& notemsg);
     static int exists (const json& j, const std::string & a){return j.find(a) != j.end();}
     static std::string add_str (const json& j, const std::string & a);
-    static uint32_t add_number (const json& j, const std::string & a);
+    static uint64_t add_number (const json& j, const std::string & a);
     static std::string add_bool (const json& j, const std::string & a);
   };
 }
